autosectask/requirements.txt
docx_mailmerge==0.5.0
openpyxl==3.0.7
pandas==1.4.3
paramiko==2.7.2
xlrd==2.0.1
xlutils==2.0.0
Pillow==8.3.1